requirements.txt
black==22.1.0
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
Pygments==2.11.2
python-dotenv==0.19.2
rich==11.2.0
snakeviz==2.1.1
tabulate==0.8.9
tomli==2.0.1
tornado==6.1
tqdm==4.63.0
